+++ devel/webpack_indexes/min.wepsim_node-end.js
@@ -1,4 +1,4 @@
 
-export * from '../../src/wepsim_nodejs/wepsim_node_core.js';
-export * from '../../src/wepsim_nodejs/wepsim_node_action.js';
+export * from '../../src_ts/src/wepsim_nodejs/wepsim_node_core.js';
+export * from '../../src_ts/src/wepsim_nodejs/wepsim_node_action.js';